refactor(entity): 修改记录实体中的考勤状态字段类型
- 将 Record 结构体中的 attendance_status 字段从 i64 类型改为 i8 类型
- 将 RecordTable 结构体中的 attendance_status 字段从 i64 类型改为 i8 类型
- 将相关联的结构体中的 attendance_status 字段从 i64 类型改为 i8 类型
- 保持数据序列化和反序列化的兼容性配置不变
- 优化内存占用，attendance_status 字段只需存储小范围整数值
- 确保数据库映射和业务逻辑处理的一致性

+++ backend/src/common/entity/record.rs
@@ -12,7 +12,7 @@ use ts_rs::TS;
 pub struct Record {
     pub id: Option<i64>,
     pub student_id: i64,
-    pub attendance_status: i64,
+    pub attendance_status: i8,
     pub remark: Option<String>,
 
     #[serde(deserialize_with = "crate::util::serde_utils::deserialize_timestamp_from_millisecond_i64")]
@@ -43,7 +43,7 @@ impl Record {
 pub struct RecordTable {
     pub id: i64,
     pub student_id: i64,
-    pub attendance_status: i64,
+    pub attendance_status: i8,
     pub remark: Option<String>,
 
     #[serde(deserialize_with = "crate::util::serde_utils::deserialize_timestamp_from_millisecond_i64")]
@@ -81,7 +81,7 @@ pub struct RollcallRecord {
     pub student_id: i64,
     pub student_no: String,
     pub name: String,
-    pub attendance_status: i64,
+    pub attendance_status: i8,
     pub remark: Option<String>,
 
     #[serde(deserialize_with = "crate::util::serde_utils::deserialize_timestamp_from_millisecond_i64")]
